Moved showErrorAndLog to VSCodeFunctions
